Trying to get paths in the right order so as to prevent "reversing" the clip mask
Want to replace this logic with the following: When generating a curve, if it reaches an "end" that doesn't touch another curve, finish as normal, then backtrack round that same curve, trying to find somewhere to branch off a new curve from. If that doesn't find one, go back to the previous curve, and so on, until all (contiguous) curves have been checked.

That should ensure that all curves in a set are done in the right order. But how do we ensure nested sets are done in the right order, to sort "holes" etc?